Refactor RecipeTile
